requirements.txt
feedparser==6.0.11
pyyaml==6.0.2
jinja2==3.1.6
markupsafe==3.0.2
python-dateutil==2.9.0
requests==2.32.3
